Replace caches w/ suspend cache
This CL replaces cache usages in Cache.kt and Inflight.kt with the new
suspend cache.

Cache.kt replacement is straightforward.

Inflight.kt replacement requires some behavior definition of how  behaves.
One option is to completely remove Inflight since it may not make much sense.
For fresh get, the behavior when there is another ongoing request is active is
not very clear. One option is to ignore that request since it might've happend
 data might have changed. Another behavior is to return the result of
ongoing request. Maybe Inflight is meant to resolve that edge case, I'm not sure.

Current implementation works in a way that it ignores any ongoing requests BUT
tries to make an attempt to prevent any further get()'s from creating a new
request