Parameterize StoreWithParserTest and make it also test pipeline impl
Added more overloads to TestStoreBuilder which separates a parser that is
added after fetch w/ parser that is added before value is returned. It is
ugly but ignoring since that class is temporary.

I've also SimplePersisterAsFlowable which converts a simple Persister that
does not support observability to one that supports via write and clear.

To properly distinguish the two, I've changed Persister's reader API to return
Flow<T?> instead of Flow<T>. Without letting it to return null, it is not
possible to implement a  since stream() never ends if it supports
observability.