Get rid of unused scope in NoOpPersister
This PR fixes an issue in NoOpPersister where it was using a CoroutineScope w/o
really needing one as there wasn't any actual deferred work going on. As a side
effect of that, it was flaking the StoreTest.

After this change, the test does not flake aynmore (tried by setting test to be re-run
until failure